gralloc_ump: initial commit
From:

http://www.malideveloper.com/developer-resources/drivers/open-source-mali-gpus-android-gralloc-module.php

This is an Android Gralloc module which allocates memory from the
Unified Memory Profider (UMP) device driver. The Gralloc module is
used by Android's composition engine in order to allocate and manage
memory which is suitable for 2D and 3D graphics usage. The Mali driver
stack is able to use UMP memory directly, resulting in zero-copy
operations when compositing.

Signed-off-by: Angus Ainslie <angus.ainslie@linaro.org>
Signed-off-by: Zach Pfeffer <zach.pfeffer@linaro.org>